I added the bootstrap MSBuild tasks to the MSBuild script.
I created versions of the ParseVersionNumber and GenerateVersionInfo
MSBuild tasks to the MSBuild script as inline tasks in order to generate
the VersionInfo.cs file for each build of the project. I added the
BuildNumber output to the ParseVersionNumber task and added code to
extract the first decimal number in the build version component as the
build number for the .NET version number.