Update MSBuild Extensions project
I updated the MSBuild Extensions project. I switched to Visual Studio
2013. I removed the use of the GenerateVersionInfo task to use the
WriteCodeFragment MSBuild task instead.

I removed the strong name key and signing from the project. The
assembly is now unsigned. I evaluated the pros and cons of strong
naming and decided that strong naming was not beneficial for this
project.